requirements.txt
GetOldTweets3==0.0.11
pandas==1.0.3
pysqlite3==0.4.2
python-dotenv==0.12.0
Flask==1.1.1
slackclient==2.5.0
slackeventsapi==2.1.0
